Number Aware Concatenation with Commas

Concatenates character vector(s) into a single string with each value separated by a comma.

cn(x, one, some = one, none = some, lots = none, bracket = "",
  nlots = 10)

Arguments

x
A vector of length n that can be coerced to character.

one
The string to return if n = 1

some
The string to return if n is in 2, 3, ..., nlots - 1

none
The string to return if n = 0

lots
The string to return if n >= nlots

nlots
A count of the number of values to consider to be lots.

See also

Examples

cn(1:3, "the %n elements are: %c")

#> [1] "the 3 elements are: 1, 2, 3"

fun <- function(x) {
 cn(x,
  one = "there is 1 element: %c",
  some = "there are %n elements: %c",
  none = "there are %n elements",
  )
 }
 fun("3")

#> [1] "there is 1 element: 3"

 fun(character(0))

#> [1] "there are 0 elements"

 fun(3:2)

#> [1] "there are 2 elements: 3, 2"

 fun(1:10)

#> [1] "there are 10 elements"